modules/main/dashboard/requirements.txt
flask
flask-cors
pandas
requests
flask-talisman
flask-httpauth
python-dotenv
gunicorn
